+++ lib/Doctrine/ODM/PHPCR/Mapping/ClassMetadata.php
@@ -1322,6 +1322,17 @@ class ClassMetadata implements ClassMetadataInterface
 
     /**
      * {@inheritDoc}
+     * @deprecated use getFieldMapping instead
+     */
+    public function getField($fieldName)
+    {
+        @trigger_error(__METHOD__.'() is deprecated since 1.1.4 and will be removed in 2.0, use getFieldMapping() instead.', E_USER_DEPRECATED);
+
+        return $this->getFieldMapping($fieldName);
+    }
+
+    /**
+     * {@inheritDoc}
      */
     public function hasAssociation($fieldName)
     {